Add OnlinePIN tests
